% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dltam-package.R
\docType{package}
\name{dltam-package}
\alias{dltam-package}
\title{dltam: Dynamic Local Trend Association Measure (DLTAM)}
\description{
Implementation of the dynamic local trend association measure (DLTAM) for time series, proposed by García-López, F. J., Batyrshin, I., and Gelbukh, A., in the paper "Dynamic Local Trend Associations in Analysis of Comovements of Financial Time Series" \doi{10.1007/978-3-319-67137-6_20}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://adamoshen.github.io/dltam}
  \item \url{https://github.com/adamoshen/dltam}
  \item Report bugs at \url{https://github.com/adamoshen/dltam/issues}
}

}
\author{
\strong{Maintainer}: Adam Shen \email{adamshen1@gmail.com}

}
\keyword{internal}
